+++ R/delta_BD.R
@@ -4,8 +4,8 @@ lin_interp = function(df, BD_min, BD_max, n=20){
   stopifnot(!is.null(df$Count))
 
   # linear interpolation function
-  BDs = df$Buoyant_density %>% as.Num
-  lin_fun = approxfun(x=BDs, y=as.Num(df$Count))
+  BDs = as.numeric(as.character(df$Buoyant_density))
+  lin_fun = stats::approxfun(x=BDs, y=as.numeric(as.character(df$Count)))
 
   # BDs (x) for interplation of abundances (y)
   BD_x = seq(BD_min, BD_max, length.out=n)
@@ -61,6 +61,8 @@ lin_interp = function(df, BD_min, BD_max, n=20){
 #' data(physeq_S2D2_l)
 #' # just selecting 1 treatment-control comparison
 #' physeq = physeq_S2D2_l[[1]]
+#'
+#' \dontrun{
 #' # calculating delta_BD
 #' df = delta_BD(physeq, control_expr='Substrate=="12C-Con"')
 #' head(df)
@@ -69,6 +71,7 @@ lin_interp = function(df, BD_min, BD_max, n=20){
 #' data(physeq_rep3)
 #' df = delta_BD(physeq_rep3, control_expr='Treatment=="12C-Con"')
 #' head(df)
+#' }
 #'
 delta_BD = function(physeq, control_expr, n=20, BD_min=NULL, BD_max=NULL){
   # atom excess
@@ -79,13 +82,14 @@ delta_BD = function(physeq, control_expr, n=20, BD_min=NULL, BD_max=NULL){
 
   # total sum scaling
   df_OTU = df_OTU %>%
-    dplyr::group_by(SAMPLE_JOIN) %>%
-    dplyr::mutate(Count = Count / sum(Count),
-                  Count = ifelse(is.na(Count), 0, Count)) %>%
+    dplyr::group_by_("SAMPLE_JOIN") %>%
+    dplyr::mutate_(Count = "as.numeric(as.character(Count))",
+                   Count = "Count / sum(Count)",
+                   Count = "ifelse(is.na(Count), 0, Count)") %>%
     dplyr::ungroup()
 
   # BD min/max
-  df_OTU$Buoyant_density = df_OTU$Buoyant_density %>% as.Num
+  df_OTU$Buoyant_density = as.numeric(as.character(df_OTU$Buoyant_density))
   if(is.null(BD_min)){
     BD_min = df_OTU$Buoyant_density %>% min
   }
@@ -94,24 +98,27 @@ delta_BD = function(physeq, control_expr, n=20, BD_min=NULL, BD_max=NULL){
   }
 
   # calculating BD shift
+  ## params for standard-eval
+  nest_cols = c('SAMPLE_JOIN', 'Count', 'Buoyant_density')
+  dots = list(~lapply(data, lin_interp, n=n, BD_min=BD_min, BD_max=BD_max))
+  dots = stats::setNames(dots, "data")
+  ## calculation
   df_OTU = df_OTU %>%
     # linear interpolation for each OTU in each gradient
-    dplyr::group_by(IS_CONTROL, OTU) %>%
-    tidyr::nest() %>%
-    dplyr::mutate(data = lapply(data, lin_interp,
-                                n=n,
-                                BD_min=BD_min,
-                                BD_max=BD_max)) %>%
-    tidyr::unnest(Count_interp = data %>% purrr::map(function(x) x)) %>%
+    dplyr::group_by_("IS_CONTROL", "OTU") %>%
+    tidyr::nest_(key_col='data',
+                 nest_cols=nest_cols) %>%
+    dplyr::mutate_(.dots=dots) %>%
+    tidyr::unnest_(unnest_cols='data') %>%
     # center of mass
-    dplyr::group_by(IS_CONTROL, OTU) %>%
-    dplyr::summarize(center_of_mass = weighted.mean(x=Buoyant_density,
-                                             w=Count_interp)) %>%
+    dplyr::group_by_("IS_CONTROL", "OTU") %>%
+    dplyr::summarize_(center_of_mass = "stats::weighted.mean(x=Buoyant_density,
+                                                             w=Count_interp)") %>%
     # delta BD
-    dplyr::group_by(OTU) %>%
-    dplyr::mutate(IS_CONTROL = ifelse(IS_CONTROL==TRUE, 'CM_control', 'CM_treatment')) %>%
-    tidyr::spread(IS_CONTROL, center_of_mass) %>%
-    dplyr::mutate(delta_BD = CM_treatment - CM_control) %>%
+    dplyr::group_by_("OTU") %>%
+    dplyr::mutate_(IS_CONTROL="ifelse(IS_CONTROL==TRUE, 'CM_control', 'CM_treatment')") %>%
+    tidyr::spread_("IS_CONTROL", "center_of_mass") %>%
+    dplyr::mutate_(delta_BD="CM_treatment - CM_control") %>%
     dplyr::ungroup()
 
   return(df_OTU)