+++ tests/testthat/test-BD_ordinations.R
@@ -13,6 +13,8 @@ test_that('Beta diversity from a list of phyloseq objects (parallel)',{
 
 
 test_that('Make a data.frame for ordination plotting (parallel)',{
+  skip_on_cran()
+
   # params for subseting
   doParallel::registerDoParallel(2)
   physeq_l_d = physeq_list_betaDiv(physeq_S2D2_l, parallel=TRUE)
@@ -24,6 +26,8 @@ test_that('Make a data.frame for ordination plotting (parallel)',{
 
 
 test_that('Plots created from phyloseq object',{
+  skip_on_cran()
+
   # params for subseting
   params = get_treatment_params(physeq_S2D2, c('Substrate', 'Day'))
   expect_is(params, 'data.frame')
@@ -52,6 +56,8 @@ test_that('Plots created from phyloseq object',{
 
 
 test_that('Plot comparing all',{
+  skip_on_cran()
+
   # params for subseting
   params = get_treatment_params(physeq_S2D2, c('Substrate', 'Day'))
   expect_is(params, 'data.frame')