+++ tests/testthat/test-BD_shift.R
@@ -1,3 +1,5 @@
+# test_file('~/dev/HTSSIP/tests/testthat/test-BD_shift.R')
+
 test_that('Percent overlap is working',{
   x = perc_overlap(0, 1, 0, 0.5)
   expect_equal(x, 50)
@@ -5,6 +7,18 @@ test_that('Percent overlap is working',{
   expect_equal(x, 100)
   x = perc_overlap(0, 0.5, 0.5, 1)
   expect_equal(x, 0)
+})
+
+test_that('parse_dist is working',{
+  skip_on_cran()
+  # calculating beta-diversity values
+  physeq_S2D2_d = phyloseq::distance(physeq_S2D2,
+                             method='unifrac',
+                             weighted=TRUE,
+                             fast=TRUE,
+                             normalized=FALSE)
+  physeq_S2D2_d = parse_dist(physeq_S2D2_d)
+  expect_is(physeq_S2D2_d, 'data.frame')
 })
 
 
@@ -25,15 +39,17 @@ expect_wmean = function(wmean){
 }
 
 test_that('BD_shift runs w/ default',{
+  skip_on_cran()
+
   ## basic call
   data(physeq_S2D2_l)
 
   # dataset 1
-  wmean = BD_shift(physeq_S2D2_l[[1]])
+  wmean = BD_shift(physeq_S2D2_l[[1]], nperm=3)
   expect_wmean(wmean)
 
   # dataset 2
-  wmean = BD_shift(physeq_S2D2_l[[2]])
+  wmean = BD_shift(physeq_S2D2_l[[2]], nperm=3)
   expect_wmean(wmean)
 
   # ggplot
@@ -42,12 +58,24 @@ test_that('BD_shift runs w/ default',{
   expect_is(p, 'ggplot')
 })
 
+test_that('BD_shift runs with differing permutation methods',{
+  skip_on_cran()
+  data(physeq_S2D2_l)
+
+  wmean = BD_shift(physeq_S2D2_l[[1]], nperm=3, perm_method='treatment')
+  expect_wmean(wmean)
+  wmean = BD_shift(physeq_S2D2_l[[1]], nperm=3, perm_method='adjacent')
+  expect_wmean(wmean)
+  wmean = BD_shift(physeq_S2D2_l[[1]], nperm=3, perm_method='overlap')
+  expect_wmean(wmean)
+})
+
 test_that('BD_shift runs w/ Bray-Curtis',{
-  ## basic call
+  skip_on_cran()
   data(physeq_S2D2_l)
 
   # dataset 1
-  wmean = BD_shift(physeq_S2D2_l[[1]], method='bray')
+  wmean = BD_shift(physeq_S2D2_l[[1]], method='bray', nperm=5)
   expect_wmean(wmean)
 
   # ggplot