+++ NEWS.md
@@ -1,3 +1,40 @@
+# HTSSIP 1.3.0
+
+* added heavy_SIP() function for conducting historically used SIP analysis methods
+* altered and expanded the permutation methods used for BD_shift()
+
+# HTSSIP 1.2.0
+
+* added permutation test to BD_shift()
+* update function documentation
+* fixed a couple of small bugs
+
+# HTSSIP 1.1.1
+
+* small documentation updates and bug fixes
+
+
+# HTSSIP 1.1.0
+
+* fixed BD ordination plotting standard evaluation bug
+
+
+# HTSSIP 1.0.3 
+
+* CRAN release
+
+
+# HTSSIP 1.0.2
+
+* Fixed CRAN namesspace warnings/notes
+* Reduced run time of examples, tests, and vignettes
+
+
+# HTSSIP 1.0.1
+
+* Reduced run time of examples, tests, and vignettes
+
+
 # HTSSIP 1.0.0
 
 * Examples for exported functions.